refacto: move Pathfinder to temper-components and create PathfinderSearch component
