+++ src/components/src/lib.rs
@@ -11,6 +11,7 @@ pub mod last_chunk_pos;
 pub mod last_synced_position;
 pub mod metadata;
 pub mod physical;
+pub mod pathfinder;
 pub mod spawn;
 
 // Interaction components re-exports
