+++ src/game_systems/src/pathfinding/src/lib.rs
@@ -3,4 +3,5 @@ mod cost;
 pub mod pathfinder;
 
 pub use astar::{Path, find_path};
-pub use pathfinder::{Pathfinder, pos_to_block, tick_pathfinder};
+pub use pathfinder::{PathfinderSearch, pos_to_block, tick_pathfinder};
+pub use temper_components::pathfinder::Pathfinder;
